🎨 Add test code about slice
Combine test code reducer and dispatch thunk actions
